requirements.txt
requests
beautifulsoup4
nltk
sentence-transformers
faiss-cpu
transformers
torch
numpy
